test_droplevel (#32272)
