+++ pandas/tests/series/test_alter_axes.py
@@ -152,16 +152,3 @@ class TestSeriesAlterAxes:
         for axis in [2, "foo"]:
             with pytest.raises(ValueError, match="No axis named"):
                 s.set_axis(list("abcd"), axis=axis, inplace=False)
-
-    def test_droplevel(self):
-        # GH20342
-        ser = Series([1, 2, 3, 4])
-        ser.index = MultiIndex.from_arrays(
-            [(1, 2, 3, 4), (5, 6, 7, 8)], names=["a", "b"]
-        )
-        expected = ser.reset_index("b", drop=True)
-        result = ser.droplevel("b", axis="index")
-        tm.assert_series_equal(result, expected)
-        # test that droplevel raises ValueError on axis != 0
-        with pytest.raises(ValueError):
-            ser.droplevel(1, axis="columns")
